+++ Changelog.md
@@ -1,3 +1,1 @@
-- Updated Ore textures
-- Fixed Diaphragms and Sieve Assemblies being opaque
-- Fixed BlockEntity Renderers
+- Added missing recipes